start refactoring registrie to make more sense to the whole shabam
