+++ src/server/extensions/middleware/registry.ts
@@ -1,7 +1,6 @@
 import type { RequestMiddleware } from "../../types";
-import { debug } from "../../utils/logger";
-
-const middlewares = new Map<string, RequestMiddleware>();
+import { pluginsDir } from "../../utils/paths";
+import { createRegistry } from "../registry-factory";
 
 function isRequestMiddleware(val: unknown): val is RequestMiddleware {
   if (typeof val !== "object" || val === null) return false;
@@ -13,56 +12,23 @@ function isRequestMiddleware(val: unknown): val is RequestMiddleware {
   );
 }
 
+const registry = createRegistry<RequestMiddleware>({
+  dirs: () => [{ dir: pluginsDir(), source: "plugin" }],
+  match: (mod) => {
+    const m = mod.middleware ?? (mod.default as Record<string, unknown>)?.middleware;
+    return isRequestMiddleware(m) ? m : null;
+  },
+  debugTag: "middleware",
+});
+
 export async function initMiddlewareRegistry(): Promise<void> {
-  const { readdir, stat } = await import("fs/promises");
-  const { join } = await import("path");
-  const { pathToFileURL } = await import("url");
-  const { pluginsDir } = await import("../../utils/paths");
-  const pluginDir = pluginsDir();
-  middlewares.clear();
-
-  try {
-    const entries = await readdir(pluginDir);
-    for (const entry of entries) {
-      const entryPath = join(pluginDir, entry);
-      const entryStat = await stat(entryPath).catch(() => null);
-      if (!entryStat?.isDirectory()) continue;
-
-      let indexFile: string | undefined;
-      for (const f of ["index.js", "index.ts", "index.mjs", "index.cjs"]) {
-        const s = await stat(join(entryPath, f)).catch(() => null);
-        if (s?.isFile()) {
-          indexFile = f;
-          break;
-        }
-      }
-      if (!indexFile) continue;
-
-      try {
-        const fullPath = join(entryPath, indexFile);
-        const url = pathToFileURL(fullPath).href;
-        const mod = await import(url);
-        const m = mod.middleware ?? mod.default?.middleware;
-        if (!m || !isRequestMiddleware(m)) continue;
-        middlewares.set(m.id, m);
-      } catch (err) {
-        debug(
-          "middleware",
-          `Failed to load middleware from plugin: ${entry}`,
-          err,
-        );
-      }
-    }
-  } catch (err) {
-    debug("middleware", "Failed to read plugin directory", err);
-  }
+  await registry.init();
 }
 
 export function getMiddleware(id: string): RequestMiddleware | null {
-  return middlewares.get(id) ?? null;
+  return registry.items().find((m) => m.id === id) ?? null;
 }
 
 export async function reloadMiddlewareRegistry(): Promise<void> {
-  middlewares.clear();
-  await initMiddlewareRegistry();
+  await registry.reload();
 }
