+++ src/server/extensions/plugin-routes/registry.ts
@@ -1,7 +1,11 @@
 import type { PluginRoute } from "../../types";
-import { debug } from "../../utils/logger";
+import { pluginsDir } from "../../utils/paths";
+import { createRegistry } from "../registry-factory";
 
-const pluginRoutes = new Map<string, PluginRoute[]>();
+interface RouteEntry {
+  pluginId: string;
+  routes: PluginRoute[];
+}
 
 function isPluginRoute(val: unknown): val is PluginRoute {
   if (typeof val !== "object" || val === null) return false;
@@ -14,66 +18,30 @@ function isPluginRoute(val: unknown): val is PluginRoute {
   );
 }
 
-function isPluginRouteArray(val: unknown): val is PluginRoute[] {
-  return Array.isArray(val) && val.every(isPluginRoute);
-}
-
 function normalizePath(p: string): string {
   const s = p.trim().replace(/^\/+/, "").replace(/\/+$/, "") || "";
   return s ? `/${s}` : "/";
 }
 
+const registry = createRegistry<RouteEntry>({
+  dirs: () => [{ dir: pluginsDir(), source: "plugin" }],
+  match: (mod) => {
+    const routes = mod.routes ?? (mod.default as Record<string, unknown>)?.routes;
+    if (!Array.isArray(routes) || !(routes as unknown[]).every(isPluginRoute) || routes.length === 0) return null;
+    return { pluginId: "", routes: (routes as PluginRoute[]).map((r) => ({ ...r, path: normalizePath(r.path) })) };
+  },
+  onLoad: async (entry, { folderName }) => {
+    entry.pluginId = folderName;
+  },
+  debugTag: "plugin-routes",
+});
+
 export async function initPluginRoutes(): Promise<void> {
-  const { readdir, stat } = await import("fs/promises");
-  const { join } = await import("path");
-  const { pathToFileURL } = await import("url");
-  const { pluginsDir } = await import("../../utils/paths");
-  const pluginDir = pluginsDir();
-  pluginRoutes.clear();
-
-  try {
-    const entries = await readdir(pluginDir);
-    for (const entry of entries) {
-      const entryPath = join(pluginDir, entry);
-      const entryStat = await stat(entryPath).catch(() => null);
-      if (!entryStat?.isDirectory()) continue;
-
-      let indexFile: string | undefined;
-      for (const f of ["index.js", "index.ts", "index.mjs", "index.cjs"]) {
-        const s = await stat(join(entryPath, f)).catch(() => null);
-        if (s?.isFile()) {
-          indexFile = f;
-          break;
-        }
-      }
-      if (!indexFile) continue;
-
-      try {
-        const fullPath = join(entryPath, indexFile);
-        const url = pathToFileURL(fullPath).href;
-        const mod = await import(url);
-        const routes = mod.routes ?? mod.default?.routes;
-        if (!isPluginRouteArray(routes) || routes.length === 0) continue;
-        const normalized = routes.map((r) => ({
-          ...r,
-          path: normalizePath(r.path),
-        }));
-        pluginRoutes.set(entry, normalized);
-      } catch (err) {
-        debug(
-          "plugin-routes",
-          `Failed to load routes from plugin: ${entry}`,
-          err,
-        );
-      }
-    }
-  } catch (err) {
-    debug("plugin-routes", "Failed to read plugin directory", err);
-  }
+  await registry.init();
 }
 
 export function getPluginRoutes(pluginId: string): PluginRoute[] {
-  return [...(pluginRoutes.get(pluginId) ?? [])];
+  return [...(registry.items().find((e) => e.pluginId === pluginId)?.routes ?? [])];
 }
 
 export function findPluginRoute(
@@ -81,17 +49,13 @@ export function findPluginRoute(
   method: string,
   path: string,
 ): PluginRoute | null {
-  const routes = pluginRoutes.get(pluginId);
-  if (!routes) return null;
+  const entry = registry.items().find((e) => e.pluginId === pluginId);
+  if (!entry) return null;
   const normalized = path.replace(/^\/+/, "").replace(/\/+$/, "") || "";
   const want = normalized ? `/${normalized}` : "/";
-  return (
-    routes.find((r) => r.method === method.toLowerCase() && r.path === want) ??
-    null
-  );
+  return entry.routes.find((r) => r.method === method.toLowerCase() && r.path === want) ?? null;
 }
 
 export async function reloadPluginRoutes(): Promise<void> {
-  pluginRoutes.clear();
-  await initPluginRoutes();
+  await registry.reload();
 }
